Work round ruby's half assed Tempfile class which, because for some reason it is implemented by delegating to File (by inheriting from DelegateClass(File)) rather than by inheriting from it, does not actually look like a File, and hence does not look like an IO (in that it does not respond to kind_of? IO calls).
In ruby 1.8.5 this doesn't matter as REXML checks that the source
it has been given responds to the methods it wants (using respond_to?
which is handled properly by DelegateClass) but in 1.8.4 which we
are running on rails2 REXML uses kind_of? to see if the source is
one it can use.