Allow to not specify a password when signing up with openID
If the password field has a length of 0 and we have a valid
openID, then disable password login by generating a random
unknown password.

This still needs some warnings to indicate that not all services
such as trac, the forum or some editors are supported with
openid only via oauth.