Change misleading wording on copyright page
Previously it said that you may distribute derived works "under the
same or similar license", now it just says "under the same".

The former form is derived from the cc-by-sa summary:

    http://creativecommons.org/licenses/by-sa/2.0/

But the legal code itself says:

    You may distribute, publicly display, publicly perform, or publicly
    digitally perform a Derivative Work only under the terms of this
    License, a later version of this License with the same License
    Elements as this License, or a Creative Commons iCommons license that
    contains the same License Elements as this
    License (e.g. Attribution-ShareAlike 2.0 Japan)

The intent is obviously to refer to other creative commons
licenses. This makes sense in the context of the creative commons
website. but on our site the reader could infer that he's able to
distribute the map data under similar licenses, such as the GPL, GFDL,
Apache 2 and others.