Switch to using oauth-plugin as a gem
Use the oauth-plugin as a gem instead of carrying a local copy.

This also required that commit eff06faf46c8f0cc4d92a16f6b2daa22eabb4893 be
reverted as it changed the plugin. I don't think the cosmetic improvement
from that patch is worth the hassle of carrying local changes to the plugin.