Try and patch up poor browser language selections
The HTTP RFC says that a language range specified by a browser only
matches a resource whose language tag is the same, or longer. This
means that a browser language range of de-DE will not match a resource
with a language tag of de.

Because of this browsers should always send the generic range, such
as de, as a secondary choice after a more specific range like de-DE.

Some browers don't this however, so if we don't get a language match
we try and patch up the list of language ranges by inserting the
more generic ranges.