Don't linkify comments in the changeset list
It's confusing to make links in comments active in the changeset list
as the rest of the comment is an active link to the changeset.

We do want to make them active in the changeset view however, so add
linkification to the comment there.

Fixes https://trac.openstreetmap.org/ticket/5061