Update key and data browser for new road colours
Update colours for main roads to match new style and replace
references to trunk, primary and secondary in the key with a
generic reference to main roads as trunk, primary and secondary
here have OSM specific meaning.

Also change road icons in the data browser to be the same colours
as the revised colours in the standard map style. The top and bottom
lines for motorway, trunk and primary are in the casing colours,
secondary and tertiary are 50% intensity of casing colour. Tertiary
is made deeper.