Make the AMF getway_old call parse times as UTC
Although it's not entirely clear how to interpret a timestamp from
a client the production servers have effectively been parsing them
as UTC because they run in the UTC time zone all year.

Using Time.zone to do the parse forces it to happen in the rails time
zone rather than the system time zone, which defaults to UTC as well.

Fixes #1688